Use unique packages for terminal interface implementations
When compiling a native image, we get the following error:
> java.lang.IncompatibleClassChangeError: com.github.ajalt.mordant.terminal.terminalinterface.WinKernel32Lib and com.github.ajalt.mordant.terminal.terminalinterface.WinKernel32Lib$EventUnion disagree on InnerClasses attribute

There were multiple classes with the same package + name
