+++ mordant-jvm-ffm/src/jvmMain/kotlin/com/github/ajalt/mordant/terminal/terminalinterface/ffm/TerminalInterface.ffm.linux.kt
@@ -1,11 +1,12 @@
-package com.github.ajalt.mordant.terminal.terminalinterface
+package com.github.ajalt.mordant.terminal.terminalinterface.ffm
 
 import com.github.ajalt.mordant.rendering.Size
+import com.github.ajalt.mordant.terminal.terminalinterface.TerminalInterfaceJvmPosix
 import java.lang.foreign.Arena
 import java.lang.foreign.MemorySegment
 
 @Suppress("ClassName", "PropertyName", "SpellCheckingInspection")
-private class MacosCLibrary {
+private class LinuxCLibrary {
     class winsize(override val segment: MemorySegment) : StructAccessor {
         object Layout : StructLayout() {
             val ws_row by shortField()
@@ -24,11 +25,16 @@ private class MacosCLibrary {
 
     class termios(override val segment: MemorySegment) : StructAccessor {
         object Layout : StructLayout() {
-            val c_iflag by longField()
-            val c_oflag by longField()
-            val c_cflag by longField()
-            val c_lflag by longField()
+            val c_iflag by intField()
+            val c_oflag by intField()
+            val c_cflag by intField()
+            val c_lflag by intField()
+            val c_line by byteField()
             val c_cc by arrayField(32)
+            @Suppress("unused")
+            val padding by paddingField(3)
+            val c_ispeed by intField()
+            val c_ospeed by intField()
         }
 
         constructor(arena: Arena) : this(arena.allocate(Layout.layout))
@@ -37,7 +43,10 @@ private class MacosCLibrary {
         var c_oflag by Layout.c_oflag
         var c_cflag by Layout.c_cflag
         var c_lflag by Layout.c_lflag
+        var c_line by Layout.c_line
         val c_cc by Layout.c_cc
+        var c_ispeed by Layout.c_ispeed
+        var c_ospeed by Layout.c_ospeed
     }
 
     object MethodHandles : MethodHandlesHolder() {
@@ -64,19 +73,19 @@ private class MacosCLibrary {
     }
 }
 
-internal class TerminalInterfaceFfmMacos : TerminalInterfaceJvmPosix() {
-    override val termiosConstants: TermiosConstants get() = MacosTermiosConstants
+internal class TerminalInterfaceFfmLinux : TerminalInterfaceJvmPosix() {
+    override val termiosConstants: TermiosConstants get() = LinuxTermiosConstants
 
     private companion object {
         const val TIOCGWINSZ = 0x00005413
         const val TCSADRAIN: Int = 0x1
     }
 
-    private val libC = MacosCLibrary()
+    private val libC = LinuxCLibrary()
     override fun isatty(fd: Int): Boolean = libC.isatty(fd)
 
     override fun getTerminalSize(): Size? = Arena.ofConfined().use { arena ->
-        val size = MacosCLibrary.winsize(arena)
+        val size = LinuxCLibrary.winsize(arena)
         if (!libC.ioctl(STDIN_FILENO, TIOCGWINSZ, size.segment)) {
             null
         } else {
@@ -85,7 +94,7 @@ internal class TerminalInterfaceFfmMacos : TerminalInterfaceJvmPosix() {
     }
 
     override fun getStdinTermios(): Termios = Arena.ofConfined().use { arena ->
-        val termios = MacosCLibrary.termios(arena)
+        val termios = LinuxCLibrary.termios(arena)
         if (!libC.tcgetattr(STDIN_FILENO, termios)) {
             throw RuntimeException("failed to read terminal settings")
         }
@@ -99,14 +108,14 @@ internal class TerminalInterfaceFfmMacos : TerminalInterfaceJvmPosix() {
     }
 
     override fun setStdinTermios(termios: Termios): Unit = Arena.ofConfined().use { arena ->
-        val nativeTermios = MacosCLibrary.termios(arena)
+        val nativeTermios = LinuxCLibrary.termios(arena)
         if (!libC.tcgetattr(STDIN_FILENO, nativeTermios)) {
             throw RuntimeException("failed to update terminal settings")
         }
-        nativeTermios.c_iflag = termios.iflag.toLong()
-        nativeTermios.c_oflag = termios.oflag.toLong()
-        nativeTermios.c_cflag = termios.cflag.toLong()
-        nativeTermios.c_lflag = termios.lflag.toLong()
+        nativeTermios.c_iflag = termios.iflag.toInt()
+        nativeTermios.c_oflag = termios.oflag.toInt()
+        nativeTermios.c_cflag = termios.cflag.toInt()
+        nativeTermios.c_lflag = termios.lflag.toInt()
         nativeTermios.c_cc.copyFrom(MemorySegment.ofArray(termios.cc))
         libC.tcsetattr(STDIN_FILENO, TCSADRAIN, nativeTermios)
     }
