+++ mordant-jvm-ffm/src/jvmMain/resources/META-INF/services/com.github.ajalt.mordant.terminal.TerminalInterfaceProvider
@@ -1,1 +1,1 @@
-com.github.ajalt.mordant.terminal.terminalinterface.TerminalInterfaceProviderFfm
+com.github.ajalt.mordant.terminal.terminalinterface.ffm.TerminalInterfaceProviderFfm
