+++ mordant-jvm-graal-ffi/src/jvmMain/resources/META-INF/services/com.github.ajalt.mordant.terminal.TerminalInterfaceProvider
@@ -1,1 +1,1 @@
-com.github.ajalt.mordant.terminal.terminalinterface.TerminalInterfaceProviderNativeImage
+com.github.ajalt.mordant.terminal.terminalinterface.nativeimage.TerminalInterfaceProviderNativeImage
