+++ mordant-jvm-jna/api/mordant-jvm-jna.api
@@ -1,4 +1,4 @@
-public final class com/github/ajalt/mordant/terminal/terminalinterface/TerminalInterfaceProviderJna : com/github/ajalt/mordant/terminal/TerminalInterfaceProvider {
+public final class com/github/ajalt/mordant/terminal/terminalinterface/jna/TerminalInterfaceProviderJna : com/github/ajalt/mordant/terminal/TerminalInterfaceProvider {
 	public fun <init> ()V
 	public fun load ()Lcom/github/ajalt/mordant/terminal/TerminalInterface;
 }
