+++ mordant-jvm-jna/src/jvmMain/resources/META-INF/services/com.github.ajalt.mordant.terminal.TerminalInterfaceProvider
@@ -1,1 +1,1 @@
-com.github.ajalt.mordant.terminal.terminalinterface.TerminalInterfaceProviderJna
+com.github.ajalt.mordant.terminal.terminalinterface.jna.TerminalInterfaceProviderJna
